src/requirements.txt
aiogram==3.3.0
aiohttp==3.9.1
redis==5.0.1
python-dotenv==1.0.0
PyGithub==2.1.1
